10215-keyboard-only.cypress.spec.js keyboard only test is now passing.
